% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.nmadas}
\alias{fit.nmadas}
\title{fit}
\usage{
fit.nmadas(nma.model = NULL, data, S.ID, T.ID, Comparator = "NA",
  tp = NULL, fn = NULL, tn = NULL, fp = NULL, cores = 3, chains = 3,
  iter = 6000, warmup = 1000, thin = 10, ...)
}
\arguments{
\item{nma.model}{A model written in the stan format from \link{nmamodel}. If the model is not specified,
a hierachical beta-binomial model with frank copula is fitted.}

\item{data}{A data-frame with no missing values containg TP, TN, FP, FN, SID and TID.}

\item{tp}{A string indicating the name of the columnt with the true positives.}

\item{fn}{A string indicating the name of the columnt with the false negatives.}

\item{tn}{A string indicating the name of the columnt with the true negatives.}

\item{fp}{A string indicating the name of the columnt with the false positives.}

\item{cores}{A positive numeric values specifying the number of cores to use to execute parallel sampling. When the hardware has more at least 4 cores,
the default is 3 cores and otherwise 1 core.}

\item{chains}{A positive numeric value specifying the number of chains, default is 3.}

\item{iter}{A positive numeric value specifying the number of iterations per chain. The default is 6000.}

\item{warmup}{A positive numeric value (<iter) specifying the number of iterations to be discarded(burn-in/warm-up). The default is 1000.}

\item{thin}{A positive numeric value specifying the interval in which the samples are stored. The default is 10.}

\item{...}{Other optional parameters as specified in \link[rstan]{stan}.}

\item{SID}{A string indicating the name of the column with the study identifier.}

\item{TID}{A string indicating the name of the column with the test identifier.}
}
\value{
An object of nmadasfit class.
}
\description{
Fit a NMA model to the data.
}
\examples{
\dontrun{
data(demodata)

fit1 <- fit(S.ID='study',
		   T.ID = 'Test',
		   tp = 'TP',
		   tn = 'TN',
		   fp = 'FP',
		   fn = 'FN',
            data = demodata,
            iter = 6000,
            warmup = 2000,
            thin = 5,
            seed = 3)

modelcode <- nmamodel.nmadas(copula = "fgm", marginals = "beta")

fit2 <- fit(nma.model = modelcode,
		   S.ID='study',
		   T.ID = 'Test',
		   tp = 'TP',
		   tn = 'TN',
		   fp = 'FP',
		   fn = 'FN',
            data = demodata,
            iter = 6000,
            warmup = 2000,
            thin = 5,
            seed = 3)
}

}
\references{
{Agresti A (2002). Categorical Data Analysis. John Wiley & Sons, Inc.}

{Clayton DG (1978). A model for Association in Bivariate Life Tables and its Application in
Epidemiological Studies of Familial Tendency in Chronic Disease Incidence. Biometrika,65(1), 141-151.}

{Frank MJ (1979). On The Simultaneous Associativity of F(x, y) and x + y - F(x, y). Aequationes Mathematicae, pp. 194-226.}

{Farlie DGJ (1960). The Performance of Some Correlation Coefficients for a General Bivariate
Distribution. Biometrika, 47, 307-323.}

{Gumbel EJ (1960). Bivariate Exponential Distributions. Journal of the American Statistical Association, 55, 698-707.}

{Meyer C (2013). The Bivariate Normal Copula. Communications in Statistics - Theory and Methods, 42(13), 2402-2422.}

{Morgenstern D (1956). Einfache Beispiele Zweidimensionaler Verteilungen. Mitteilungsblatt furMathematische Statistik, 8, 23 - 235.}

{Sklar A (1959). Fonctions de Repartition a n Dimensions et Leurs Marges. Publications de l'Institut de Statistique de L'Universite de Paris, 8, 229-231.}
}
\author{
Victoria N Nyaga <victoria.nyaga@outlook.com>
}
