% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,nmadasmodel-method}
\title{A function to fit the model.}
\usage{
fit(object, ...)

\S4method{fit}{nmadasmodel}(object, data, S.ID, T.ID, Comparator = "NA",
  tp = NULL, fn = NULL, tn = NULL, fp = NULL, cores = 3, chains = 3,
  iter = 6000, warmup = 1000, thin = 10, ...)
}
\arguments{
\item{...}{Other optional parameters as specified in \link[rstan]{stan}.}

\item{data}{A data-frame with no missing values containg TP, TN, FP, FN, SID and TID.}

\item{Comparator}{The name of the comparator test when relative sensitivity and specificity are required. By default the first test as
arranged alphabetically is the comparator.}

\item{tp}{A string indicating the name of the columnt with the true positives.}

\item{fn}{A string indicating the name of the columnt with the false negatives.}

\item{tn}{A string indicating the name of the columnt with the true negatives.}

\item{fp}{A string indicating the name of the columnt with the false positives.}

\item{cores}{A positive numeric values specifying the number of cores to use to execute parallel sampling. When the hardware has more at least 4 cores,
the default is 3 cores and otherwise 1 core.}

\item{chains}{A positive numeric value specifying the number of chains, default is 3.}

\item{iter}{A positive numeric value specifying the number of iterations per chain. The default is 6000.}

\item{warmup}{A positive numeric value (<iter) specifying the number of iterations to be discarded(burn-in/warm-up). The default is 1000.}

\item{thin}{A positive numeric value specifying the interval in which the samples are stored. The default is 10.}

\item{nma.model}{A model written in the stan format from \link{nmadasmodel}. If the model is not specified,
a hierachical beta-binomial model with frank copula is fitted.}

\item{SID}{A string indicating the name of the column with the study identifier.}

\item{TID}{A string indicating the name of the column with the test identifier.}
}
\value{
An object of nmadasfit class.
}
\description{
Fit a NMA model to the data.

A function to fit the model.
}
