% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkplot.R
\name{networkplot.nmadas}
\alias{networkplot.nmadas}
\title{Network plot}
\usage{
networkplot.nmadas(data, S.ID, T.ID, alphabetic = FALSE, edge.col = "grey",
  node.col = "orange", ...)
}
\arguments{
\item{data}{A data.frame with atleast two variables i.e study and test identifier.}

\item{...}{additional options. See \link[pcnetmeta]{nma.networkplot} for more details.}

\item{SID}{A text indicating the name of the variable that identifies the different studies.}

\item{TID}{A text indicating the name of the variable that identifies the different tests.}
}
\description{
Draw a network plot.
}
\examples{
data(demodata)
networkplot(data = demodata,
            SID = "study",
            TID = "Test")
}
\author{
Victoria N Nyaga
}
