% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{nmadasdata-class}
\alias{nmadasdata-class}
\title{Class nmadasdata}
\description{
A nmadasdata class in the NMADAS package.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data-frame with no missing values containg TP, TN, FP, FN, study and test names.}

\item{\code{S.ID}}{Study identifier}

\item{\code{T.ID}}{Test identifier}

\item{\code{labels}}{A vector with test labels.}
}}

\seealso{
\link{cdtamodel}

Other nmadas: \code{\link{nmadasfit-class}},
  \code{\link{nmadasmodel-class}}
}
\author{
Victoria N Nyaga \email{victoria.nyaga@outlook.com}
}
