% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{nmadasmodel-class}
\alias{nmadasmodel-class}
\title{Class nmadasmodel}
\description{
A nmadasmodel class in the NMADAS package.
}
\section{Slots}{

\describe{
\item{\code{model}}{Model specification in a character format.}
}}

\seealso{
\link{nmadasmodel}

Other nmadas: \code{\link{nmadasdata-class}},
  \code{\link{nmadasfit-class}}
}
\author{
Victoria N Nyaga \email{victoria.nyaga@outlook.com}
}
