% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCode.R
\name{nmamodel.nmadas}
\alias{nmamodel.nmadas}
\title{nmamodel}
\usage{
nmamodel.nmadas(marginals = "beta", copula = "frank", p.omega = NULL,
  fullcov = FALSE, prior.lmu = "normal(0, 5)",
  prior.tau = "cauchy(0, 2.5)", prior.sigma = " cauchy(0, 2.5)",
  prior.rho = "lkj_corr(2.0)")
}
\arguments{
\item{marginals}{Use normal marginals on the logit transformed sensitivity and specificity or the beta marginals.
When marginals = 'normal' the following model is fitted:
\deqn{Y_{ijk} ~ bin(\pi_{ijk}, N_{ijk})}
\deqn{logit(\pi_{ijk}) = \mu_{jk} + \eta_{ij} + \delta_{ijk}}
\deqn{(\eta_{i1}, \eta_{i2})' ~ N_2(0, \Sigma)}
\deqn{\Sigma[j,j] = \sigma[j]^2, \Sigma[1,2] = \Sigma[2,1] = \rho*\sigma[1]*\sigma[2]}
\deqn{\delta_{ijk} ~ N(0, \tau_{jk})}}

\item{copula}{Name of the copula function used to model the correlation between sensitivity and specificty. This
requires that marginals = 'beta' be specificied.
This is a string naming the copula function. The choices are "fgm", "frank", "gauss", "c90" and "c270".}

\item{p.omega}{The prior distribution of the \eqn{\omega} parameters. This prior distribution depend on the
specified copula. The defualt is \code{"uniform(-1, 1)"} for the gaussian and fgm copula,
\code{"normal(0, 5)"} for the frank copula, ad \code{"cauchy(0, 2.5)"} for the c90 and c270 copula.}

\item{fullcov}{Logical for full (TRUE) or reduced (default) variance-covariance matrix. The reduction simplifies
the variance-covariance matrix by specifying that
\deqn{\delta_{ijk} ~ N(0, \tau_{j})}.}

\item{prior.lmu}{A text specifying the prior distribution for \eqn{\mu}  parameters. The default is \code{"normal(0, 5)"}.}

\item{prior.tau}{A text specifying the prior distribution for \eqn{\tau}  parameters. The default is \code{"cauchy(0, 2.5)"}.}

\item{prior.sigma}{A text specifying the prior distribution for \eqn{\sigma}  parameters. The default is \code{"cauchy(0, 2.5)"}.}

\item{prior.rho}{A text specifying the prior distribution for \eqn{\rho}  parameter. The default is \code{"lkj(2.0)"}.
For more details on specifications of prior distributions see \href{http://mc-stan.org/documentation/}{Stan documentation}.
When marginals = 'beta' the following model is fitted:
\deqn{Y_{ijk} ~ bin(\pi_{ijk}, N_{ijk})}
\deqn{\pi_{i1k}, \pi_{i2k} ~ f(\pi_{i1k})*f(\pi_{i2k})*copula(F(\pi_{i1k}), F(\pi_{i2k}), \omega_k)}
where f and F are the probability density and cumulative distribution function of a beta distribution with parameters
\deqn{\alpha_{jk}} and \deqn{\beta_{jk}} specified as follows:
\deqn{\alpha_{jk} = \mu_{jk}*\frac{1- \theta_j*\delta_{jk}}{\theta_j*\delta_{jk}}}
\deqn{\beta_{jk} = (1 - \mu_{jk})*\frac{1- \theta_j*\delta_{jk}}{\theta_j*\delta_{jk}}}
Here \deqn{\mu_{jk}} is the mean sensitivity \code{j = 1} and specificity \code{j = 2},
 \deqn{\omega_k} captures the correlation between sensitivity and specificity in test k,
 \deqn{\theta_j} captures the common overdispersion among the sensitivities \code{j = 1} and specificities \code{j = 2},
 and \deqn{\delta_jk} captures the test specific extra variability.
 The hyper parameters \eqn{\mu}, \eqn{\theta} and \eqn{\delta} are given beta/uniform priors since they are in the (0,1) interval.
 The prior distribution of \eqn{\omega} depends on the copula.}
}
\value{
An object of nmamodel class.
}
\description{
Specify the copula based bivariate beta-binomial or alternatively logistic-binomial distribution to fit to the diagnostic data.
}
\examples{
data(demodata)
model1 <-  nmamodel()

model2 <- nmamodel(copula = 'fgm')

model3 <-  nmamodel(marginals = 'normal')

}
\references{
{Agresti A (2002). Categorical Data Analysis. John Wiley & Sons, Inc.}

{Clayton DG (1978). A model for Association in Bivariate Life Tables and its Application in
Epidemiological Studies of Familial Tendency in Chronic Disease Incidence. Biometrika,65(1), 141-151.}

{Frank MJ (1979). On The Simultaneous Associativity of F(x, y) and x + y - F(x, y). Aequationes Mathematicae, pp. 194-226.}

{Farlie DGJ (1960). The Performance of Some Correlation Coefficients for a General Bivariate
Distribution. Biometrika, 47, 307-323.}

{Gumbel EJ (1960). Bivariate Exponential Distributions. Journal of the American Statistical Association, 55, 698-707.}

{Meyer C (2013). The Bivariate Normal Copula. Communications in Statistics - Theory and Methods, 42(13), 2402-2422.}

{Morgenstern D (1956). Einfache Beispiele Zweidimensionaler Verteilungen. Mitteilungsblatt furMathematische Statistik, 8, 23 - 235.}

{Sklar A (1959). Fonctions de Repartition a n Dimensions et Leurs Marges. Publications de l'Institut de Statistique de L'Universite de Paris, 8, 229-231.}
}
\author{
Victoria N Nyaga
}
