% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{forestplot}
\alias{forestplot}
\alias{forestplot,nmadasfit-method}
\title{A function to produce forest plots.}
\usage{
forestplot(object, vlinecolour = "blue", textsize = 4,
  pointcolour = "grey70", pointsize = 2, dp = 2,
  textlabel = "Mean [95\% CI]", dodgewidth = 1, RR = TRUE, ...)

\S4method{forestplot}{nmadasfit}(object, vlinecolour = "blue", textsize = 4,
  pointcolour = "grey70", pointsize = 2, dp = 2,
  textlabel = "Mean [95\% CI]", dodgewidth = 1, RR = TRUE, ...)
}
\arguments{
\item{object}{A nmadasfit object from \link{fit}.}
}
\description{
A function to produce forest plots.
}
