% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,nmadasmodel-method}
\alias{show,nmadasmodel-method}
\alias{show,nmadasfit-method}
\title{A function to print the model.}
\usage{
\S4method{show}{nmadasmodel}(object)

\S4method{show}{nmadasfit}(object)
}
\arguments{
\item{object}{A nmadasmodel object returned by \link{nmadasmodel} function.}
}
\description{
A function to print the model.

A function to print the results.
}
\keyword{internal}
