% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot.nmadasfit}
\alias{traceplot.nmadasfit}
\title{Traceplot}
\usage{
traceplot.nmadasfit(x, pars = c("MU"), colourset = "mix-red-brightblue",
  ...)
}
\arguments{
\item{x}{An nmadas object from \link{fit}.}

\item{...}{additional options. See \link[base]{array} for more details.}
}
\value{
A ggplot trace plot of the parameters of the models mean structure.
}
\description{
Trace plot using ggplot2.
}
\examples{
\dontrun{
data(demodata)

frank <- nmadasmodel()

fit1 <- fit(
        nma.model = frank,
        S.ID='study',
		   T.ID = 'Test',
		   tp = 'TP',
		   tn = 'TN',
		   fp = 'FP',
		   fn = 'FN',
            data = demodata,
            iter = 6000,
            warmup = 2000,
            thin = 5,
            seed = 3)

traceplot(fit1)
}
}
\author{
Victoria N Nyaga
}
