% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{traceplot}
\alias{traceplot}
\alias{traceplot,nmadasfit-method}
\title{A function to produce traceplots.}
\usage{
traceplot(object, pars = c("MU"), colourset = "mix-red-brightblue", ...)

\S4method{traceplot}{nmadasfit}(object, pars = c("MU"),
  colourset = "mix-red-brightblue", ...)
}
\arguments{
\item{object}{A cdtafit object from \link{fit}}

\item{...}{Extra optional arguments as defined in \link[rstan]{stan_trace}.}
}
\description{
A function to produce traceplots.

A function to produce traceplots.
}
